# src/Api/Responses/ClientGetAllResponse.cs
using QaCoders_Net.Entities;

namespace QaCoders_Net.Responses;

public class ClientGetAllResponse
{
    public int ClientId { get; set; }
    public string Name { get; set; }
    public DateTime CreatedDate { get; set; }

    public static explicit operator ClientGetAllResponse(Client client)
    {
        return new()
        {
            ClientId = client.ClientId,
            Name = client.Name,
            CreatedDate = client.CreateDate
        };
    }
}