# src/Api/swagger-generated.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QaCoders-Net",
    "version": "1.0"
  },
  "paths": {
    "/clients": {
      "get": {
        "tags": [
          "Client"
        ],
        "operationId": "GetClientAll",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientGetAllResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientGetAllResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientGetAllResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Client"
        ],
        "operationId": "CreateClient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QaCoders_Net.Requests.ClientCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QaCoders_Net.Requests.ClientCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QaCoders_Net.Requests.ClientCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientCreateResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientCreateResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/clients/{clientId}": {
      "get": {
        "tags": [
          "Client"
        ],
        "operationId": "GetClientByIdAsync",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientGetByIdResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientGetByIdResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientGetByIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Client"
        ],
        "operationId": "PatchClient",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QaCoders_Net.Requests.ClientPatchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QaCoders_Net.Requests.ClientPatchRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QaCoders_Net.Requests.ClientPatchRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientPatchResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientPatchResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientPatchResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "get": {
        "tags": [
          "Order"
        ],
        "operationId": "GetOrderAll",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QaCoders_Net.Responses.OrderGetByIdResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QaCoders_Net.Responses.OrderGetByIdResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QaCoders_Net.Responses.OrderGetByIdResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Order"
        ],
        "operationId": "CreateOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QaCoders_Net.Requests.OrderCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QaCoders_Net.Requests.OrderCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QaCoders_Net.Requests.OrderCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.OrderCreateResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.OrderCreateResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.OrderCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{orderId}": {
      "get": {
        "tags": [
          "Order"
        ],
        "operationId": "GetOrderByIdAsync",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.OrderGetByIdResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.OrderGetByIdResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.OrderGetByIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "QaCoders_Net.Requests.ClientCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Requests.ClientPatchRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Requests.OrderCreateRequest": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QaCoders_Net.Requests.OrderProductRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Requests.OrderProductRequest": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "QaCoders_Net.Responses.ClientCreateResponse": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Responses.ClientGetAllResponse": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Responses.ClientGetByIdResponse": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Responses.ClientPatchResponse": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Responses.ClientResponse": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Responses.ErrorMessageResponse": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Responses.ErrorResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QaCoders_Net.Responses.ErrorMessageResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Responses.OrderCreateResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "client": {
            "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientResponse"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QaCoders_Net.Responses.OrderProductResponse"
            },
            "nullable": true
          },
          "total": {
            "type": "number",
            "format": "decimal",
            "readOnly": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Responses.OrderGetByIdResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "client": {
            "$ref": "#/components/schemas/QaCoders_Net.Responses.ClientResponse"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QaCoders_Net.Responses.OrderProductResponse"
            },
            "nullable": true
          },
          "total": {
            "type": "number",
            "format": "decimal",
            "readOnly": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "QaCoders_Net.Responses.OrderProductResponse": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/QaCoders_Net.Responses.ProductResponse"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "number",
            "format": "decimal",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "QaCoders_Net.Responses.ProductResponse": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number",
            "format": "decimal"
          }
        },
        "additionalProperties": false
      }
    }
  }
}